+++ SleepModel/Sense-Bridging-Header.h
@@ -66,3 +66,5 @@
 #import "HEMMathUtil.h"
 #import "HEMSettingsHeaderFooterView.h"
 #import "HEMSettingsStoryboard.h"
+#import "HEMListPresenter.h"
+#import "HEMListItemCell.h"