+++ SleepModel/UIBarButtonItem+HEMNav.h
@@ -8,10 +8,12 @@
 
 #import <UIKit/UIKit.h>
 
+NS_ASSUME_NONNULL_BEGIN
+
 @interface UIBarButtonItem (HEMNav)
 
-+ (UIBarButtonItem*)cancelItemWithTitle:(NSString*)title
-                                  image:(UIImage*)image
++ (UIBarButtonItem*)cancelItemWithTitle:(nullable NSString*)title
+                                  image:(nullable UIImage*)image
                                  target:(id)target
                                  action:(SEL)action;
 
@@ -22,3 +24,5 @@
 + (UIBarButtonItem*)infoButtonWithTarget:(id)target action:(SEL)action;
 
 @end
+
+NS_ASSUME_NONNULL_END